split out testutil package, move method
