+++ state_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/hashcloak/katzenmint-pki/s11n"
+	"github.com/hashcloak/katzenmint-pki/testutil"
 	"github.com/katzenpost/core/crypto/ecdh"
 	"github.com/katzenpost/core/crypto/eddsa"
 	"github.com/katzenpost/core/crypto/rand"
@@ -22,7 +23,7 @@ const testEpoch = 1
 
 func TestUpdateDescriptor(t *testing.T) {
 	require := require.New(t)
-	desc, _ := CreateTestDescriptor(require, 1, pki.LayerProvider, testEpoch)
+	desc, _ := testutil.CreateTestDescriptor(require, 1, pki.LayerProvider, testEpoch)
 	rawDesc := make([]byte, 10)
 	enc := codec.NewEncoderBytes(&rawDesc, jsonHandle)
 	if err := enc.Encode(desc); err != nil {
@@ -60,7 +61,7 @@ func TestUpdateDocument(t *testing.T) {
 	require := require.New(t)
 
 	// Create, validate and deserialize document.
-	_, sDoc := CreateTestDocument(require, testEpoch)
+	_, sDoc := testutil.CreateTestDocument(require, testEpoch)
 	dDoc, err := s11n.VerifyAndParseDocument([]byte(sDoc))
 	if err != nil {
 		t.Fatalf("Failed to VerifyAndParseDocument document: %+v\n", err)
